update: stable code for writing herc model data back to DTS.

+++ ES2Core/src/main/java/org/hercworks/core/data/file/dts/bsp/TSBSPGroupNode.java
@@ -22,7 +22,7 @@ public class TSBSPGroupNode {
 		this.back = back;
 	}
 
-	public int getCoeff() {
+	public short getCoeff() {
 		return coeff;
 	}
